vexpress: move initialization routines to plat-common/
main_init_primary(), main_init_secondary(), main_cpu_on_handler() are moved
to a separate directory so that they may be re-used by the HiKey platform.
