+++ core/arch/arm/plat-common/entry_a32.S
