+++ core/arch/arm/plat-common/sub.mk
@@ -1,1 +1,5 @@
 srcs-y += tee_entry.c
+srcs-$(CFG_COMMON_INIT) += init.c
+ifeq ($(CFG_COMMON_INIT),y)
+srcs-$(CFG_ARM32_core) += entry_a32.S
+endif
