+++ core/arch/arm/plat-vexpress/conf.mk
@@ -40,6 +40,7 @@ CFG_PL011 := y
 CFG_GIC := y
 CFG_HWSUPP_MEM_PERM_PXN := y
 CFG_WITH_STACK_CANARIES := y
+CFG_COMMON_INIT := y
 
 ifeq ($(PLATFORM_FLAVOR),juno)
 CFG_CRYPTO_SHA256_ARM32_CE ?= y
