+++ core/arch/arm/plat-vexpress/sub.mk
@@ -1,5 +1,4 @@
 global-incdirs-y += .
-srcs-$(CFG_ARM32_core) += entry_a32.S
 srcs-y += main.c
 srcs-y += tee_common_otp.c
 srcs-y += core_bootcfg.c
